Added new Output file with novel functionality and and updated rest of simulation files

+++ Simulation.h
@@ -71,5 +71,5 @@ struct KineticResult {
 struct ForceResult{
     float accumulation;
     float potential;
-    vector<int> virial;
+    vector<float> virial;
 };
